Unify money as BigDecimal and quantity as Integer (#13)
Money fields become BigDecimal so financial arithmetic is exact:
Product.price, OrderDetails.price/subtotal, Orders.orderTotal (and the
/products/TotalPrice endpoint). Quantity fields are plain Integer:
Product.stock changes from BigInteger to Integer (OrderDetails.qty was
already Integer). Line-item and order-total math now uses BigDecimal.add/
multiply instead of int/BigInteger.

A V4 Flyway migration alters the Postgres columns (money -> numeric(38,2),
stock -> integer) so ddl-auto=validate still passes. H2 derives the types
from the entities. Tests updated to the new types.

Verified end-to-end: a 19.99 x 3 line item yields subtotal 59.97 and
orderTotal 59.97 with no rounding drift, stock decrements as an int, and
Flyway V1..V4 apply cleanly with validation passing on a real Postgres.

+++ backend/src/main/java/com/example/marketing/shop/domain/OrderDetails.java
@@ -6,6 +6,8 @@ import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Setter
@@ -13,8 +15,8 @@ public class OrderDetails{
     @Id
     private String id;
     private Integer Qty;
-    private Integer Price;
-    private Integer Subtotal;
+    private BigDecimal Price;
+    private BigDecimal Subtotal;
 
     @ManyToOne
     @JoinColumn(name="Order_id")
